refactor testcode
