+++ src/test/java/seedu/address/model/person/predicate/FixedAddressPredicateTest.java
@@ -1,0 +1,4 @@
+package seedu.address.model.person.predicate;
+
+public class FixedAddressPredicateTest {
+}
