+++ src/test/java/seedu/address/model/person/predicate/ShowSpecifiedPersonPredicateTest.java
@@ -1,0 +1,4 @@
+package seedu.address.model.person.predicate;
+
+public class ShowSpecifiedPersonPredicateTest {
+}
